Grande mise a jour

+++ Tables/HISTORIQUES_VENTE.sql
@@ -2,7 +2,7 @@
 
 CREATE TABLE historiques_vente
 (
-    id_client NUMBER(20) CONSTRAINT id_hv_pk NOT NULL,
+    id_client INT CONSTRAINT id_hv_pk NOT NULL,
     date_vente DATE NOT NULL,
     id_vente VARCHAR2(20) NOT NULL
 );
